add event listener at check box and bin
find out a problem , It does not form a new box downwards,
but is crowded in the same box. This is because each click only generates three divs:
text, check, and bin. These three elements are all located in the task-container
that has been defined in HTML and laid out as flex. Because it is flex,
each click will continuously use the flex feature to continuously add horizontal space.
try to fix:
 can remove the task-container from the HTML
and dynamically generate it in js with each click.